fix(web): keep workspace flag nav filters client-safe
Move navigation filtering out of workspace-flags so the client sidebar
does not import the database module.
