+++ apps/hyperlocalise-web/src/lib/flags/workspace-flag-navigation.ts
@@ -1,0 +1,57 @@
+import type { NavigationGroup, NavigationItem } from "@/components/app-shell/navigation-config";
+
+import {
+  WORKSPACE_AUTOMATIONS_FLAG,
+  WORKSPACE_ISSUES_FLAG,
+  WORKSPACE_KNOWLEDGE_FLAG,
+  WORKSPACE_VISUAL_MOCK_FLAG,
+  type WorkspaceFeatureFlagState,
+} from "./workos-flag-entities";
+
+function workspaceFlagEnabledByKey(flags: WorkspaceFeatureFlagState): Record<string, boolean> {
+  return {
+    [WORKSPACE_AUTOMATIONS_FLAG]: flags.automations,
+    [WORKSPACE_KNOWLEDGE_FLAG]: flags.knowledge,
+    [WORKSPACE_VISUAL_MOCK_FLAG]: flags.visualMock,
+    [WORKSPACE_ISSUES_FLAG]: flags.issues,
+  };
+}
+
+function isNavigationItemEnabledByWorkspaceFlags(
+  item: Pick<NavigationItem, "featureFlagKey">,
+  enabledByKey: Record<string, boolean>,
+) {
+  if (!item.featureFlagKey) {
+    return true;
+  }
+
+  if (!(item.featureFlagKey in enabledByKey)) {
+    return true;
+  }
+
+  return enabledByKey[item.featureFlagKey] ?? false;
+}
+
+export function filterNavigationItemsByWorkspaceFlags(
+  items: readonly NavigationItem[],
+  flags: WorkspaceFeatureFlagState,
+): readonly NavigationItem[] {
+  const enabledByKey = workspaceFlagEnabledByKey(flags);
+  return items.filter((item) => isNavigationItemEnabledByWorkspaceFlags(item, enabledByKey));
+}
+
+export function filterNavigationByWorkspaceFlags(
+  groups: readonly NavigationGroup[],
+  flags: WorkspaceFeatureFlagState,
+): readonly NavigationGroup[] {
+  const enabledByKey = workspaceFlagEnabledByKey(flags);
+
+  return groups
+    .map((group) => ({
+      ...group,
+      items: group.items.filter((item) =>
+        isNavigationItemEnabledByWorkspaceFlags(item, enabledByKey),
+      ),
+    }))
+    .filter((group) => group.items.length > 0);
+}
